+++ R/theme_e61.R
@@ -50,7 +50,7 @@ theme_e61 <- function(y_top = TRUE,
                       panel_borders = TRUE,
                       background = "white",
                       base_size = 10,
-                      base_family = "Arial",
+                      base_family = "ArialMT",
                       base_line_size = points_to_mm(0.75),
                       base_rect_size = points_to_mm(1)
                       ) {